+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.6 (unreleased)
+
+### Summary
+
+- [#173](https://github.com/awslabs/amazon-s3-find-and-forget/pull/173): Show
+  column types and hierarchy in the front-end during Data Mapper creation
+- [#173](https://github.com/awslabs/amazon-s3-find-and-forget/pull/173): Add
+  support for char, smallint, tinyint, double, float
+
 ## v0.5
 
 ### Summary